Fixed tabs (now spaces), reformatted tests into a test_master and began on regex tests
