+++ comment_scan/commander.rb
@@ -2,147 +2,147 @@ require 'time'
 require './db'
 
 class Commander
-	attr_reader :chatter, :BOT_NAMES, :start_time
+    attr_reader :chatter, :BOT_NAMES, :start_time
 
-	def initialize(chatter, bot_names)
-		@chatter = chatter
-		@BOT_NAMES = bot_names
+    def initialize(chatter, bot_names)
+        @chatter = chatter
+        @BOT_NAMES = bot_names
 
-		@basic_commands = Hash.new()
-		@HQ_commands = Hash.new()
+        @basic_commands = Hash.new()
+        @HQ_commands = Hash.new()
 
-		@basic_commands["!!/whoami"] = method(:whoami)
-		@basic_commands["!!/alive"] = method(:alive)
-		@basic_commands["!!/on"] = method(:on)
-		@basic_commands["!!/off"] = method(:off)
-		@basic_commands["!!/mode"] = method(:mode)
-		@basic_commands["!!/help"] = method(:help)
-		@basic_commands["!!/notify"] = method(:notify)
-		@basic_commands["!!/reports"] = method(:reports)
+        @basic_commands["!!/whoami"] = method(:whoami)
+        @basic_commands["!!/alive"] = method(:alive)
+        @basic_commands["!!/on"] = method(:on)
+        @basic_commands["!!/off"] = method(:off)
+        @basic_commands["!!/mode"] = method(:mode)
+        @basic_commands["!!/help"] = method(:help)
+        @basic_commands["!!/notify"] = method(:notify)
+        @basic_commands["!!/reports"] = method(:reports)
 
-		@HQ_commands["!!/whitelist"] = method(:whitelistuser)
-		@HQ_commands["!!/unwhitelist"] = method(:unwhitelistuser)
-		@HQ_commands["!!/whitelisted"] = method(:whitelisted)
-		@HQ_commands["!!/quota"] = method(:quota)
-		@HQ_commands["!!/uptime"] = method(:uptime)
-		@HQ_commands["!!/logsize"] = method(:logsize)
-		@HQ_commands["!!/howmany"] = method(:howmany)
-		@HQ_commands["!!/test"] = method(:test)
-		@HQ_commands["!!/howgood"] = method(:howgood)
-		@HQ_commands["!!/del"] = method(:del)
-		@HQ_commands["!!/add"] = method(:add)
-		@HQ_commands["!!/cid"] = method(:cid)
-		@HQ_commands["!!/pull"] = method(:pull)
-		@HQ_commands["!!/master"] = method(:master)
-		@HQ_commands["!!/restart"] = method(:restart)
-		@HQ_commands["!!/kill"] = method(:kill)
-		@HQ_commands["!!/rev"] = method(:rev)
-		@HQ_commands["!!/manscan"] = method(:manscan)
-		@HQ_commands["!!/ttscan"] = method(:ttscan)
-		@HQ_commands["!!/regexes"] = method(:regexes)
-		@HQ_commands["!!/regexstats"] = method(:regexstats)
+        @HQ_commands["!!/whitelist"] = method(:whitelistuser)
+        @HQ_commands["!!/unwhitelist"] = method(:unwhitelistuser)
+        @HQ_commands["!!/whitelisted"] = method(:whitelisted)
+        @HQ_commands["!!/quota"] = method(:quota)
+        @HQ_commands["!!/uptime"] = method(:uptime)
+        @HQ_commands["!!/logsize"] = method(:logsize)
+        @HQ_commands["!!/howmany"] = method(:howmany)
+        @HQ_commands["!!/test"] = method(:test)
+        @HQ_commands["!!/howgood"] = method(:howgood)
+        @HQ_commands["!!/del"] = method(:del)
+        @HQ_commands["!!/add"] = method(:add)
+        @HQ_commands["!!/cid"] = method(:cid)
+        @HQ_commands["!!/pull"] = method(:pull)
+        @HQ_commands["!!/master"] = method(:master)
+        @HQ_commands["!!/restart"] = method(:restart)
+        @HQ_commands["!!/kill"] = method(:kill)
+        @HQ_commands["!!/rev"] = method(:rev)
+        @HQ_commands["!!/manscan"] = method(:manscan)
+        @HQ_commands["!!/ttscan"] = method(:ttscan)
+        @HQ_commands["!!/regexes"] = method(:regexes)
+        @HQ_commands["!!/regexstats"] = method(:regexstats)
 
-		@start_time = Time.now
-	end
+        @start_time = Time.now
+    end
 
-	def setup_basic_commands()
-		#For each room, add each basic command
-		(@chatter.rooms + [@chatter.HQroom]).each do |room_id|
-		    @basic_commands.each do |command, action|
-				@chatter.add_command_action(room_id, command, action, [self])
-			end
-		end
-	end
+    def setup_basic_commands()
+        #For each room, add each basic command
+        (@chatter.rooms + [@chatter.HQroom]).each do |room_id|
+            @basic_commands.each do |command, action|
+                @chatter.add_command_action(room_id, command, action, [self])
+            end
+        end
+    end
 
-	def setup_HQ_commands()
-		@HQ_commands.each do |command, action|
-			@chatter.add_command_action(@chatter.HQroom, command, action, [self])
-		end
-	end
+    def setup_HQ_commands()
+        @HQ_commands.each do |command, action|
+            @chatter.add_command_action(@chatter.HQroom, command, action, [self])
+        end
+    end
 
-	def matches_bot?(botname)
-  		puts "Checking if #{botname} matches #{@BOT_NAMES}"
-  		botname.nil? || botname == '*' || @BOT_NAMES.include?(botname.downcase)
-	end
+    def matches_bot?(botname)
+        puts "Checking if #{botname} matches #{@BOT_NAMES}"
+        botname.nil? || botname == '*' || @BOT_NAMES.include?(botname.downcase)
+    end
 
-	def isHQ?(room_id)
-		room_id == @chatter.HQroom
-	end
+    def isHQ?(room_id)
+        room_id == @chatter.HQroom
+    end
 
-	def on?(room_id)
-		isHQ?(room_id) || Room.on?(room_id)
-	end
+    def on?(room_id)
+        isHQ?(room_id) || Room.on?(room_id)
+    end
 
-	#TODO: consider making some kind of check valid arg that'll return a bool and print "Bad! Arg must be in {X, Y, Z}"
-	#      since this is definitely something I do several times below
+    #TODO: consider making some kind of check valid arg that'll return a bool and print "Bad! Arg must be in {X, Y, Z}"
+    #      since this is definitely something I do several times below
 end
 
 def whoami(commander, room_id)
-	return unless commander.on?(room_id)
+    return unless commander.on?(room_id)
 
-	if commander.isHQ?(room_id) && rand(0...20) == rand(0...20)
-    	commander.chatter.say("24601")
+    if commander.isHQ?(room_id) && rand(0...20) == rand(0...20)
+        commander.chatter.say("24601")
     else
-    	commander.chatter.say("I go by #{commander.BOT_NAMES.join(" and ")}", room_id)
-	end
+        commander.chatter.say("I go by #{commander.BOT_NAMES.join(" and ")}", room_id)
+    end
 end
 
 def alive(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot) || !commander.on?(room_id)
-	commander.chatter.say("I'm alive and well :)", room_id)
+    return unless commander.matches_bot?(bot) || !commander.on?(room_id)
+    commander.chatter.say("I'm alive and well :)", room_id)
 end
 
 def on(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
-	if Room.on?(room_id)
-		commander.chatter.say("I'm already on, silly", room_id)
-	else
-		commander.chatter.say("Turning on...", room_id)
-		Room.turn_on(room_id)
-	end
+    return unless commander.matches_bot?(bot)
+    if Room.on?(room_id)
+        commander.chatter.say("I'm already on, silly", room_id)
+    else
+        commander.chatter.say("Turning on...", room_id)
+        Room.turn_on(room_id)
+    end
 end
 
 def off(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
-	if !Room.on?(room_id)
-		commander.chatter.say("I'm already off, silly", room_id)
-	else
-		commander.chatter.say("Turning off...", room_id)
-		Room.turn_off(room_id)
-	end
+    return unless commander.matches_bot?(bot)
+    if !Room.on?(room_id)
+        commander.chatter.say("I'm already off, silly", room_id)
+    else
+        commander.chatter.say("Turning off...", room_id)
+        Room.turn_off(room_id)
+    end
 end
 
 def mode(commander, room_id, bot='*')
- 	return unless commander.matches_bot?(bot) && commander.on?(room_id)
- 	if commander.isHQ?(room_id)
-    	commander.chatter.say("I'm in parent mode. I have children in rooms #{commander.chatter.rooms.map { |rid| "[#{rid}](https://chat.stackexchange.com/rooms/#{rid})"}.join(", ")}", room_id)
- 	else
-  		commander.chatter.say("I'm in child mode. My parent is in [room #{commander.chatter.HQroom}](https://chat.stackexchange.com/rooms/#{commander.chatter.HQroom})", room_id)
- 	end
+    return unless commander.matches_bot?(bot) && commander.on?(room_id)
+    if commander.isHQ?(room_id)
+        commander.chatter.say("I'm in parent mode. I have children in rooms #{commander.chatter.rooms.map { |rid| "[#{rid}](https://chat.stackexchange.com/rooms/#{rid})"}.join(", ")}", room_id)
+    else
+        commander.chatter.say("I'm in child mode. My parent is in [room #{commander.chatter.HQroom}](https://chat.stackexchange.com/rooms/#{commander.chatter.HQroom})", room_id)
+    end
 end
 
 def help(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot) && commander.on?(room_id)
-	if commander.isHQ?(room_id)
-    	commander.chatter.say(File.read('./hq_help.txt'), room_id)
-	else
+    return unless commander.matches_bot?(bot) && commander.on?(room_id)
+    if commander.isHQ?(room_id)
+        commander.chatter.say(File.read('./hq_help.txt'), room_id)
+    else
         commander.chatter.say(File.read("./help.txt"), room_id)
-	end
+    end
 end
 
 def notify(commander, room_id, bot, type, status)
- 	return unless commander.matches_bot?(bot) && commander.on?(room_id)
+    return unless commander.matches_bot?(bot) && commander.on?(room_id)
     actions = { "regex" => :regex_match,
-            	 "magic" => :magic_comment}
+                 "magic" => :magic_comment}
     if !actions.key?(type)
-    	commander.chatter.say("Type must be one of {#{actions.keys.join(", ")}}")
-    	return
+        commander.chatter.say("Type must be one of {#{actions.keys.join(", ")}}")
+        return
     end
     act = actions[type]
 
     if !["on", "off"].include? status
-    	commander.chatter.say("Status must be one of {on, off}")
-    	return
+        commander.chatter.say("Status must be one of {on, off}")
+        return
     end
     status = {"on" => true, "off" => false}[status]
 
@@ -151,50 +151,50 @@ def notify(commander, room_id, bot, type, status)
 end
 
 def reports(commander, room_id, bot='*')
- 	return unless commander.matches_bot?(bot) && !commander.on?(room_id)
+    return unless commander.matches_bot?(bot) && !commander.on?(room_id)
     room = Room.find_by(room_id: room_id)
     commander.chatter.say("regex_match: #{!!room.regex_match}\nmagic_comment: #{!!room.magic_comment}", room_id)
 end
 
 def whitelistuser(commander, room_id, bot, *uids)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     uids.each { |uid| WhitelistedUser.create(user_id: uid) }
     commander.chatter.say("Whitelisted users #{uids.join(', ')}", room_id)
 end
 
 def unwhitelistuser(commander, room_id, bot, *uids)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     uids.each { |uid| WhitelistedUser.where(user_id: uid).destroy_all }
     commander.chatter.say("Unwhitelisted users #{uids.join(', ')}", room_id)
 end
 
 def whitelisted(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     commander.chatter.say("Current whitelist: #{WhitelistedUser.all.map(&:user_id).join(', ')}", room_id)
 end
 
 def quota(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
-	#TODO: Figure out the cli situation
+    return unless commander.matches_bot?(bot)
+    #TODO: Figure out the cli situation
     #commander.chatter.say "#{cli.quota} requests remaining"
 end
 
 def uptime(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
 
-	total_secs = (Time.now - commander.start_time).to_i
-	secs = total_secs % 60; total_secs /= 60
-	mins = total_secs % 60; total_secs /= 60
-	hrs = total_secs % 24; total_secs /= 24
-	days = total_secs
+    total_secs = (Time.now - commander.start_time).to_i
+    secs = total_secs % 60; total_secs /= 60
+    mins = total_secs % 60; total_secs /= 60
+    hrs = total_secs % 24; total_secs /= 24
+    days = total_secs
 
     commander.chatter.say("Up #{days} Days, #{hrs} hours, #{mins} minutes, #{secs} seconds", room_id)
 end
 
 def logsize(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     uncompressed = to_sizes(Dir['*.log']+Dir['*.log.1']).map do |sizes|
-    	"#{sizes[:file]}: #{sizes[:size].round(2)}#{sizes[:ext]}"
+        "#{sizes[:file]}: #{sizes[:size].round(2)}#{sizes[:ext]}"
     end
     compressed = {}
     to_sizes(Dir['*.log*.gz']).each do |size|
@@ -205,29 +205,29 @@ def logsize(commander, room_id, bot='*')
 end
 
 def howmany(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     commander.chatter.say("I've scanned #{Comment.count} comments", room_id)
 end
 
 def test(commander, room_id, bot, type, *body)
-	return unless commander.matches_bot?(bot)
-	if !["q", "a"].include? type[0]
-		commander.chatter.say("Type must be one of {q, a}", room_id)
-		return
-	end
+    return unless commander.matches_bot?(bot)
+    if !["q", "a"].include? type[0]
+        commander.chatter.say("Type must be one of {q, a}", room_id)
+        return
+    end
 
-	#TODO: Get this up and running when we figure out how we're reporting
+    #TODO: Get this up and running when we figure out how we're reporting
     #say(report(type, body.join(" ")) || "Didn't match any filters")
 end
 
 def howgood(commander, room_id, bot, type, regex)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     type = 'question' if type == 'q'
     type = 'answer' if type == 'a'
-	if !["q", "a", "*"].include? type
+    if !["q", "a", "*"].include? type
          commander.chatter.say("Type must be one of {q, a, question, answer, *}", room_id)
-		return
-	end
+        return
+    end
     if type == '*'
         tps = Comment.where("tps >= ?", 1).count { |comment| %r{#{regex}}.match(comment.body_markdown.downcase) }.to_f
         fps = Comment.where("fps >= ?", 1).count { |comment| %r{#{regex}}.match(comment.body_markdown.downcase) }.to_f
@@ -239,130 +239,130 @@ def howgood(commander, room_id, bot, type, regex)
     end
 
     tp_msg = [ # Generate tp line
-     	'tp'.center(6),
-     	tps.round(0).to_s.center(11),
-     	percent_str(tps, total).center(14),
-     	percent_str(tps, Comment.where("tps >= ?", 1).count).center(15),
-     	percent_str(tps, Comment.count).center(18),
+        'tp'.center(6),
+        tps.round(0).to_s.center(11),
+        percent_str(tps, total).center(14),
+        percent_str(tps, Comment.where("tps >= ?", 1).count).center(15),
+        percent_str(tps, Comment.count).center(18),
     ].join('|')
 
     fp_msg = [ # Generate fp line
-     	'fp'.center(6),
-     	fps.round(0).to_s.center(11),
-     	percent_str(fps, total).center(14),
-     	percent_str(fps, Comment.where("fps >= ?", 1).count).center(15),
-     	percent_str(fps, Comment.count).center(18),
+        'fp'.center(6),
+        fps.round(0).to_s.center(11),
+        percent_str(fps, total).center(14),
+        percent_str(fps, Comment.where("fps >= ?", 1).count).center(15),
+        percent_str(fps, Comment.count).center(18),
     ].join('|')
 
     total_msg = [ # Generate total line
-    	'Total'.center(6),
-    	total.round(0).to_s.center(11),
-    	'-'.center(14),
-    	'-'.center(15),
-    	percent_str(total, Comment.count).center(18),
+        'Total'.center(6),
+        total.round(0).to_s.center(11),
+        '-'.center(14),
+        '-'.center(15),
+        percent_str(total, Comment.count).center(18),
     ].join('|')
 
     # Generate header line
     header = " Type | # Matched | % of Matched | % of All Type | % of ALL Comments"
 
     final_output = [ # Add 4 spaces for formatting and newlines
-    	header, '-'*68, tp_msg, fp_msg, total_msg
+        header, '-'*68, tp_msg, fp_msg, total_msg
     ].join("\n    ")
     msgs = MessageCollection::ALL_ROOMS
     msgs.push_howgood [regex, type], (commander.chatter.say("    #{final_output}", room_id))
 end
 
 def del(commander, room_id, bot, type, regex)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     if r = Regex.find_by(post_type: type[0], regex: regex)
-     	reas_id = r["reason_id"]
-     	commander.chatter.say("Destroyed #{r.regex} (post_type #{r.post_type})!", room_id) if r.destroy
+        reas_id = r["reason_id"]
+        commander.chatter.say("Destroyed #{r.regex} (post_type #{r.post_type})!", room_id) if r.destroy
 
-     	# If there are no other regexes for this reason, destroy the reason too
-     	if Regex.where(reason_id: reas_id).empty?
-    		reas = Reason.where(id: reas_id)[0]
-    		commander.chatter.say("Destroyed reason: \"#{reas["name"]}\"", room_id) if reas.destroy
-    	end
+        # If there are no other regexes for this reason, destroy the reason too
+        if Regex.where(reason_id: reas_id).empty?
+            reas = Reason.where(id: reas_id)[0]
+            commander.chatter.say("Destroyed reason: \"#{reas["name"]}\"", room_id) if reas.destroy
+        end
     else
-    	commander.chatter.say("Could not find regex to destroy", room_id)
+        commander.chatter.say("Could not find regex to destroy", room_id)
     end
 
 end
 
 def add(commander, room_id, bot, type, regex, *reason)
-	return unless commander.matches_bot?(bot)
-	if !["q", "a"].include? type
-		commands.chatter.say("Type must be one of {q, a}", room_id)
-		return
-	end
+    return unless commander.matches_bot?(bot)
+    if !["q", "a"].include? type
+        commands.chatter.say("Type must be one of {q, a}", room_id)
+        return
+    end
 
     if reason = Reason.find_or_create_by(name: reason.join(' '))
-    	begin
-     		%r{#{regex}}
-    	rescue RegexpError => e
-     		commander.chatter.say("Invalid regex: #{regex}", room_id)
-     		commander.chatter.say("    #{e}", room_id)
-    	else
-     		if r = reason.regexes.create(post_type: type[0], regex: regex)
-       			commander.chatter.say("Added regex #{r.regex} for post_type #{r.post_type} with reason '#{r.reason.name}'", room_id)
-     		end
-    	end
+        begin
+            %r{#{regex}}
+        rescue RegexpError => e
+            commander.chatter.say("Invalid regex: #{regex}", room_id)
+            commander.chatter.say("    #{e}", room_id)
+        else
+            if r = reason.regexes.create(post_type: type[0], regex: regex)
+                commander.chatter.say("Added regex #{r.regex} for post_type #{r.post_type} with reason '#{r.reason.name}'", room_id)
+            end
+        end
     end
 end
 
 def cid(commander, room_id, bot, *cids)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     c = Comment.find_by(comment_id: cid)
     if c
-    	#TODO: Fix this once we figure out how reporting comments works
-    	#report_comments(c, cli: cli, settings: settings, cb: cb, should_post_matches: false)
+        #TODO: Fix this once we figure out how reporting comments works
+        #report_comments(c, cli: cli, settings: settings, cb: cb, should_post_matches: false)
     else
-    	commander.chatter.say("Could not find comment with id #{cid}", room_id)
+        commander.chatter.say("Could not find comment with id #{cid}", room_id)
     end
 
 end
 
 def pull(commander, room_id, bot='*', num_to_post=0, update_bundle=false)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     if `git symbolic-ref --short HEAD`.chomp != "master"
-    	commander.chatter.say("Pulling is only permitted when running on the master branch. Currently on #{`git rev-parse --abbrev-ref HEAD`.chop}.", room_id)
+        commander.chatter.say("Pulling is only permitted when running on the master branch. Currently on #{`git rev-parse --abbrev-ref HEAD`.chop}.", room_id)
     else
-    	`git pull`
-    	#TODO: Fix this once we know how restart will work
-    	#restart num_to_post, bundle
+        `git pull`
+        #TODO: Fix this once we know how restart will work
+        #restart num_to_post, bundle
     end
 
 end
 
 def master(commander, room_id, bot='*', *args)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     if `git symbolic-ref --short HEAD`.chomp == "master"
-    	commander.chatter.say("I'm already on master!", room_id)
+        commander.chatter.say("I'm already on master!", room_id)
     else
-    	`git checkout master`
-    	Kernel.exec("bundle exec ruby comment_scan.rb #{args.empty? ? @post_on_startup : args[0].to_i}")
+        `git checkout master`
+        Kernel.exec("bundle exec ruby comment_scan.rb #{args.empty? ? @post_on_startup : args[0].to_i}")
     end
 end
 
 def restart(commander, room_id, bot='*', num_to_post=0, update_bundle=false)
-	return unless commander.matches_bot?(bot)
-	#TODO: Figure out how restart will work
+    return unless commander.matches_bot?(bot)
+    #TODO: Figure out how restart will work
     #restart num_to_post, bundle
 end
 
 def kill(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     `kill -9 $(cat bot.pid)`
 end
 
 def rev(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     commander.chatter.say("Currently at rev #{`git rev-parse --short HEAD`.chop} on branch #{`git rev-parse --abbrev-ref HEAD`.chop}", room_id)
 end
 
 def manscan(commander, room_id, bot='*', *cids)
-	return unless commander.matches_bot?(bot)
-	#TODO: Figure out cli
+    return unless commander.matches_bot?(bot)
+    #TODO: Figure out cli
     #c = cli.comments(args)
     #if c.empty?
       #say "No comments found for id(s) #{args.join(", ")}"
@@ -373,16 +373,16 @@ def manscan(commander, room_id, bot='*', *cids)
 end
 
 def ttscan(commander, room_id, bot='*')
-	return unless commander.matches_bot?(bot)
-	#TODO: Figure out how to get time till next scan (need to write the scanner)
+    return unless commander.matches_bot?(bot)
+    #TODO: Figure out how to get time till next scan (need to write the scanner)
     #commander.chatter.say("#{sleeptime} seconds remaning until the next scan", room_id)
 end
 
 def regexes(commander, room_id, bot='*', reason=nil)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     reasons = (reason.nil? ? Reason.all : Reason.where("name LIKE ?", "%#{reason}%")).map do |r|
-    	regexes = r.regexes.map { |regex| "- #{regex.post_type}: #{regex.regex}" }
-    	"#{r.name.gsub(/\(\@(\w*)\)/, '(*\1)')}:\n#{regexes.join("\n")}"
+        regexes = r.regexes.map { |regex| "- #{regex.post_type}: #{regex.regex}" }
+        "#{r.name.gsub(/\(\@(\w*)\)/, '(*\1)')}:\n#{regexes.join("\n")}"
     end
     reasonless_regexes = Regex.where(reason_id: nil).map { |regex| "- #{regex.post_type}: #{regex.regex}" }
     reasons << "Other Regexes:\n#{reasonless_regexes.join("\n")}"
@@ -390,15 +390,15 @@ def regexes(commander, room_id, bot='*', reason=nil)
 end
 
 def regexstats(commander, room_id, bot='*', reason=nil)
-	return unless commander.matches_bot?(bot)
+    return unless commander.matches_bot?(bot)
     #Build array of hashes for each regex containing info to build the stat output
     regexes = (reason.nil? ? Reason.all : Reason.where("name LIKE ?", "%#{reason}%")).map do |r|
-     	r.regexes.map do |regex| 
-       		tps = Comment.where("tps >= ?", 1).count { |comment| %r{#{regex.regex}}.match(comment.body_markdown.downcase) }
-       		fps = Comment.where("fps >= ?", 1).count { |comment| %r{#{regex.regex}}.match(comment.body_markdown.downcase) }
-       		{:effectivePercent => (tps + fps > 0) ? tps/(tps + fps).to_f : 0, 
-        		:tps => tps, :fps => fps, :postType => regex.post_type, :regex => regex.regex, :reason => r.name}
-     	end
+        r.regexes.map do |regex| 
+            tps = Comment.where("tps >= ?", 1).count { |comment| %r{#{regex.regex}}.match(comment.body_markdown.downcase) }
+            fps = Comment.where("fps >= ?", 1).count { |comment| %r{#{regex.regex}}.match(comment.body_markdown.downcase) }
+            {:effectivePercent => (tps + fps > 0) ? tps/(tps + fps).to_f : 0, 
+                :tps => tps, :fps => fps, :postType => regex.post_type, :regex => regex.regex, :reason => r.name}
+        end
     end
     regexes = regexes.flatten.sort_by { |regex| regex[:effectivePercent] } #Order by effectiveness
 
@@ -409,12 +409,12 @@ def regexstats(commander, room_id, bot='*', reason=nil)
 
     #Put it all together...
     commander.chatter.say regexes.map { |r| 
-     	[
-       		"".ljust(4),
-        	percent_str(r[:tps], r[:tps] + r[:fps],
+        [
+            "".ljust(4),
+            percent_str(r[:tps], r[:tps] + r[:fps],
                     precision: 1, blank_str: "n/a").ljust(percent_width),
-       		"(#{r[:tps]}/#{r[:tps] + r[:fps]})".ljust(tpfp_width),
-       		"| #{r[:regex]} (#{r[:postType]} - #{r[:reason]})"
-     	].join
+            "(#{r[:tps]}/#{r[:tps] + r[:fps]})".ljust(tpfp_width),
+            "| #{r[:regex]} (#{r[:postType]} - #{r[:reason]})"
+        ].join
     }.join("\n"), room_id
 end
