+++ comment_scan/tests/commander_tests/commander_tests.rb
@@ -1,0 +1,4 @@
+require 'test/unit'
+
+require_relative 'whitelist_tests'
+require_relative 'regex_tests'
