+++ comment_scan/tests/test_master.rb
@@ -1,0 +1,3 @@
+require 'test/unit'
+
+require_relative 'commander_tests/commander_tests'
